+++ tests/testthat/test-read-csv.R
@@ -138,7 +138,9 @@ test_that("too few or extra col_types generates warnings", {
 # read_csv2 ---------------------------------------------------------------
 
 test_that("decimal mark automatically set to ,", {
-  x <- read_csv2("x\n1,23", progress = FALSE)
+  expect_message(
+    x <- read_csv2("x\n1,23", progress = FALSE),
+    if (default_locale()$decimal_mark == ".") "decimal .*grouping .*mark" else NA)
   expect_equal(x[[1]], 1.23)
 })